changed file structure

+++ construct_square_matrix.cpp
@@ -6,7 +6,7 @@
  * input: a positive integer n and number x
  * output:  an n × n matrix with each element equal to x
  * *****************************************************/
-std::vector<std::vector<int>>  construct_square_matrix(int n, double x) {
+std::vector<std::vector<int>>  construct_square_matrix(int n, int x) {
     std::vector<std::vector<int>> rows;
     for(int i=0; i<n; i++) {
         rows.push_back({});
